feat: add wear to fuel tanks and seats (TASK-45 phase 0c)
- extract TileWearable base (wear state + capability) from TileBrokenPart
- TileBrokenPart now extends it, keeps motor render and staged drop
- fuel tanks and seats host a TileWearable (TE-only wear, no meta/render)
- register TileWearable; seats wear at 0.25x rate
- wear is captured into the rocket storage chunk and ticked on landing
