+++ src/main/java/zmaster587/advancedRocketry/tile/TileBrokenPart.java
@@ -1,98 +1,31 @@
 package zmaster587.advancedRocketry.tile;
 
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.util.EnumFacing;
-import net.minecraftforge.common.capabilities.Capability;
-import zmaster587.advancedRocketry.api.capability.CapabilityWear;
-import zmaster587.advancedRocketry.api.capability.IPartWear;
 import zmaster587.advancedRocketry.util.IBrokenPartBlock;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.Random;
 
-public class TileBrokenPart extends TileEntitySyncable implements IPartWear {
-
-    private int stage;
-    private int maxStage;
-    private float transitionProb;
-    private float[] probs;
-    private final Random rand;
+/**
+ * Wear host for rocket motors: a {@link TileWearable} that also renders the
+ * breaking overlay (motors render INVISIBLE and rely on this TESR) and drops a
+ * staged item so a worn motor keeps its wear when picked up and replaced.
+ */
+public class TileBrokenPart extends TileWearable {
 
     public TileBrokenPart() {
-        this(0, 0);
+        super();
     }
 
     public TileBrokenPart(int stage, int maxStage, float transitionProb, Random rand) {
-        this.stage = stage;
-        this.maxStage = maxStage;
-        this.rand = rand;
-        this.initProb(transitionProb);
+        super(stage, maxStage, transitionProb, rand);
     }
 
     public TileBrokenPart(int maxStage, float transitionProb, Random rand) {
-        this(0, maxStage, transitionProb, rand);
+        super(maxStage, transitionProb, rand);
     }
 
     public TileBrokenPart(int maxStage, float transitionProb) {
-        this(maxStage, transitionProb, new Random());
-    }
-
-    public void setStage(int stage) {
-        this.stage = stage;
-        this.markDirty();
-    }
-
-    public int getStage() {
-        return this.stage;
-    }
-
-    public int getMaxStage() {
-        return this.maxStage;
-    }
-
-    private void initProb(float transitionProb) {
-        this.transitionProb = transitionProb;
-        this.probs = new float[maxStage];
-
-        for (int i = 0; i < maxStage; i++) {
-            this.probs[i] = transitionProb / (float) Math.sqrt(2 * i + 1);
-        }
-    }
-
-    public boolean transition() {
-        if (stage == maxStage) {
-            return true;
-        }
-        for (int i = maxStage - 1; i >= 0; i--) {
-            if (stage == i) {
-                return false;
-            }
-            if (rand.nextFloat() < (stage + 1) * this.probs[i]) {
-                stage = i;
-                this.markDirty();
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public boolean hasCapability(@Nonnull Capability<?> capability, @Nullable EnumFacing facing) {
-        if (capability == CapabilityWear.PART_WEAR) {
-            return true;
-        }
-        return super.hasCapability(capability, facing);
-    }
-
-    @Nullable
-    @Override
-    public <T> T getCapability(@Nonnull Capability<T> capability, @Nullable EnumFacing facing) {
-        if (capability == CapabilityWear.PART_WEAR) {
-            return CapabilityWear.PART_WEAR.cast(this);
-        }
-        return super.getCapability(capability, facing);
+        super(maxStage, transitionProb);
     }
 
     @Override
@@ -102,24 +35,5 @@ public class TileBrokenPart extends TileEntitySyncable implements IPartWear {
 
     public ItemStack getDrop() {
         return ((IBrokenPartBlock) this.getBlockType()).getDropItem(world.getBlockState(pos), world, this);
-    }
-
-    @Nonnull
-    @Override
-    public NBTTagCompound writeToNBT(final NBTTagCompound compound) {
-        compound.setInteger("stage", stage);
-        compound.setInteger("maxStage", maxStage);
-        compound.setFloat("transitionProb", transitionProb);
-        return super.writeToNBT(compound);
-    }
-
-    @Override
-    public void readFromNBT(@Nonnull final NBTTagCompound compound) {
-        super.readFromNBT(compound);
-        stage = compound.getInteger("stage");
-        maxStage = compound.getInteger("maxStage");
-        transitionProb = compound.getFloat("transitionProb");
-
-        this.initProb(transitionProb);
     }
 }
